Fix context in beforeEach call (unittest)
This was happen by the difference between Jasmin(CodePen) and Jest(Motocal)

It's repored in official jest issue
Jest fixed the re-use of `this` context as bug.
(it's implicitly context binding to use `this` in test code)
so Jasmin's test code does will not work in Jest.

This patch is required in #201